@@ -1,7 +1,7 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
     defaultConfig {
         applicationId "com.dtec.helloatu"
         minSdkVersion 15
@@ -32,27 +32,31 @@ android {
 }
 
 dependencies {
-
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
     implementation 'com.android.support.constraint:constraint-layout:1.1.3'
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
-
-    implementation 'com.android.support:appcompat-v7:23.3.0'
-    implementation 'com.android.support:design:23.3.0'
-    implementation 'com.android.support:recyclerview-v7:23.3.0'
-    implementation 'com.android.support:cardview-v7:23.3.0'
+    //implementation 'com.android.support:appcompat-v7:23.3.0'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    //implementation 'com.android.support:design:23.3.0'
+    implementation 'com.google.android.material:material:1.0.0-rc01'
+    //implementation 'com.android.support:recyclerview-v7:23.3.0'
+    implementation 'androidx.recyclerview:recyclerview:1.0.0'
+    //implementation 'com.android.support:cardview-v7:23.3.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
     implementation 'com.github.bumptech.glide:glide:3.7.0'
     implementation 'com.nostra13.universalimageloader:universal-image-loader:1.9.5'
     implementation 'com.android.volley:volley:1.1.0'
     implementation 'uk.co.chrisjenx:calligraphy:2.2.0'
     implementation 'com.theartofdev.edmodo:android-image-cropper:2.3.+'
     implementation 'com.nbsp:library:1.09'
+    implementation 'com.google.android.material:material:1.0.0-rc01'
+
     implementation 'com.google.code.gson:gson:2.3.1'
     implementation 'com.itextpdf:itextg:5.5.9'
     implementation 'de.greenrobot:greendao:2.1.0'
-
-    implementation 'com.android.support:support-v4:23.4.0'
-
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.0-alpha5'
+    //implementation 'com.android.support:support-v4:23.4.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
 }
