@@ -2,7 +2,8 @@ package com.dtec.helloatu.utilities;
 
 import android.content.Context;
 import android.content.pm.PackageManager;
-import android.support.v4.app.ActivityCompat;
+
+import androidx.core.app.ActivityCompat;
 
 public class PermissionCheck {
 
